chore: rename loss_calculator to pytorch_loss_calculator
