+++ src/spira_training/shared/adapters/pytorch/model_trainer/implementations/pytorch_loss_calculator/single_loss_calculator.py
